[gtk] -Fix: zoom/pan: rounding errors
git-svn-id: svn://repo.hu/pcb-rnd/trunk@25047 1d61de99-560f-4dba-8957-1e505d42358b

+++ src_plugins/lib_gtk_common/ui_zoompan.c
@@ -206,7 +206,7 @@ void pcb_gtk_zoom_view_win_side(pcb_gtk_view_t *v, pcb_coord_t x1, pcb_coord_t y
 	}
 }
 
-void pcb_gtk_pan_view_abs(pcb_gtk_view_t *v, pcb_coord_t pcb_x, pcb_coord_t pcb_y, int widget_x, int widget_y)
+void pcb_gtk_pan_view_abs(pcb_gtk_view_t *v, pcb_coord_t pcb_x, pcb_coord_t pcb_y, double widget_x, double widget_y)
 {
 	v->x0 = pcb_round((double)SIDE_X(v, pcb_x) - (double)widget_x * v->coord_per_px);
 	v->y0 = pcb_round((double)SIDE_Y(v, pcb_y) - (double)widget_y * v->coord_per_px);
